fix buffer mismatch

+++ classix/merge_cm_win.pyx
@@ -263,7 +263,7 @@ cpdef distance_merge(double[:, :] data, list labels,
     cdef long[:] ul = np.unique(sp_cluster_labels)
     cdef Py_ssize_t nr_u = len(ul)
 
-    cdef long[:] cs = np.zeros(nr_u, dtype=int)
+    cdef long[:] cs = np.zeros(nr_u, dtype=np.int32)
     
     cdef np.ndarray[np.npy_bool, ndim=1, cast=True] cid
     cdef np.ndarray[np.int32_t, ndim=1] grp_sizes = splist.base[:, 1]
